custom/template/requirements-dev.txt
# linting and formatting
black
flake8
pylint-odoo
